Time: 0 ms (100.00%) | Memory: 7.1 MB (61.35%) - LeetSync

+++ 633-sum-of-square-numbers/sum-of-square-numbers.cpp
@@ -2,12 +2,13 @@ class Solution {
 public:
     bool judgeSquareSum(int c) {
 
-        int left  = 0;
-        int right = sqrt(c);
+        long long left  = 0;
+        long long right = sqrt(c);
 
         while(left <= right){
 
-            long long val = static_cast<long long>(left) * left + static_cast<long long>(right) * right;
+            //long long val = static_cast<long long>(left) * left + static_cast<long long>(right) * right;
+            long long val = left * left + right * right;
 
             if(val == c){
                 return true;
